Handle error when processing a full mqtt packet
When `s_process_mqtt_packet` is called for a pending packet, the result
is checked and an error is returned in case of a failure, which
eventually terminates the connection.

We have to do the same check when reading a full packet at once,
otherwise the connection will be kept open when an invalid packet is
encountered.